+++ tests/test_index/test_operation.py
@@ -1,21 +1,17 @@
-from typing import AsyncIterator
-from typing import cast
 
+from typing import AsyncIterator, cast
 import pytest
-
-from dipdup.config import DipDupConfig
-from dipdup.config import OperationHandlerConfig
-from dipdup.config import OperationHandlerOriginationPatternConfig
-from dipdup.config import OperationIndexConfig
-from dipdup.datasources.tzkt.datasource import TzktDatasource
-from dipdup.indexes.operation.fetcher import get_origination_filters
-from dipdup.indexes.operation.fetcher import get_transaction_filters
+from dipdup.config import DipDupConfig, OperationHandlerConfig, OperationHandlerOriginationPatternConfig, OperationIndexConfig
+from dipdup.datasources.datasource import Datasource
+from dipdup.enums import OperationType
+from dipdup.exceptions import FrameworkException
+from dipdup.indexes.operation.fetcher import get_origination_filters, get_transaction_filters
 from tests import CONFIGS_PATH
 from tests import tzkt_replay
 
 
 @pytest.fixture
-async def tzkt() -> AsyncIterator[TzktDatasource]:
+async def tzkt() -> AsyncIterator[Datasource]:
     async with tzkt_replay() as tzkt:
         yield tzkt
 
@@ -23,26 +19,26 @@ async def tzkt() -> AsyncIterator[TzktDatasource]:
 @pytest.fixture
 def index_config() -> OperationIndexConfig:
     config = DipDupConfig.load([CONFIGS_PATH / 'operation_filters.yml'], True)
-    config.initialize(skip_imports=True)
+    config.initialize()
     return cast(OperationIndexConfig, config.indexes['test'])
 
 
 async def test_ignored_type_filter(
-    tzkt: TzktDatasource,
+    tzkt: Datasource,
     index_config: OperationIndexConfig,
 ) -> None:
     index_config.types = ()
-    addresses, hashes = await get_origination_filters(index_config, tzkt)
+    addresses, hashes = await get_origination_filters(index_config, tzkt)  # type: ignore[arg-type]
     assert not addresses
     assert not hashes
 
-    addresses, hashes = await get_transaction_filters(index_config, tzkt)
+    addresses, hashes = await get_transaction_filters(index_config, tzkt)  # type: ignore[arg-type]
     assert not addresses
     assert not hashes
 
 
 async def test_get_origination_filters(
-    tzkt: TzktDatasource,
+    tzkt: Datasource,
     index_config: OperationIndexConfig,
 ) -> None:
     index_config.handlers = (
@@ -83,9 +79,9 @@ async def test_get_origination_filters(
             ),
         ),
     )
-    addresses, hashes = await get_origination_filters(index_config, tzkt)
-    assert not addresses
-    assert hashes == {-680664524}
+    # NOTE: Resolved earlier
+    with pytest.raises(FrameworkException):
+        await get_origination_filters(index_config, tzkt)
 
     index_config.handlers = (
         OperationHandlerConfig(
@@ -102,11 +98,13 @@ async def test_get_origination_filters(
     assert hashes == set()
 
 
-# async def test_get_transaction_filters(tzkt: TzktDatasource, index_config: OperationIndexConfig) -> None:
-#     index_config.types = (OperationType.transaction,)
-#     addresses, hashes = await get_transaction_filters(index_config, tzkt)
-#     assert filters == ({'KT1RJ6PbjHpwc3M5rw5s2Nbmefwbuwbdxton'}, {-680664524, -1585533315})
+async def test_get_transaction_filters(tzkt: Datasource, index_config: OperationIndexConfig) -> None:
+    index_config.types = (OperationType.transaction,)
+    index_config.contracts[2].code_hash = -680664524
 
-#     index_config.types = ()
-#     addresses, hashes = await get_transaction_filters(index_config, tzkt)
-#     assert filters == (set(), set())
+    filters = await get_transaction_filters(index_config, tzkt)
+    assert filters == ({'KT1RJ6PbjHpwc3M5rw5s2Nbmefwbuwbdxton'}, {-680664524, -1585533315})
+
+    index_config.types = ()
+    filters = await get_transaction_filters(index_config, tzkt)
+    assert filters == (set(), set())