+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 ### Fixed
 
 - tzkt: Fixed issue with processing rollbacks while sync is in progress.
+- tzkt: Fixed operation matching when contract code hash specified as a string.
 
 ## [6.5.11] - 2023-09-02